Move code compilation to the VM
